Refactor code and create utility function for check isOnRWS
