+++ packages/extension/src/utils/fetchRWSInfo.ts
