+++ packages/extension/src/utils/findRWSURLSets.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * External dependencies.
+ */
+import { getDomain } from 'tldts';
+
+/**
+ * Internal dependencies.
+ */
+import type { RelatedWebsiteSetType } from '../@types';
+
+const findRWSURLSets = (
+  domain: string | null,
+  rwsSets: RelatedWebsiteSetType[]
+): RelatedWebsiteSetType | undefined => {
+  return rwsSets.find((rws: RelatedWebsiteSetType) => {
+    const rwsDomains: string[] = Object.keys(rws.rationaleBySite || {}).map(
+      (_url) => getDomain(_url) || ''
+    );
+
+    if (domain === getDomain(rws.primary)) {
+      return true;
+    }
+
+    return domain ? rwsDomains.includes(domain) : false;
+  });
+};
+
+export default findRWSURLSets;
